basic_tutorial notebook suggestions (#19)
* basic_tutorial runthrough, improvement suggestions

I imposed a couple of conventions:

- column and row names in the tables shown are in the text written in
**bold**
- variable names, parameters, and code in general is formatted as `code`
- files/names in the Zarr directory are formatted as 'cellData'

* Dockerfile, tzdata

The existing tzdata solution did not work for me when building the
image, with these updates it worked for me

* Dockerfile, dask methods working

Scarf method `Datastore.plot_marker_heatmap` did not work without the
installed package fsspec

* Dockerfile, automatic scarf installation

I added automatic installation of the local scarf module. Not sure if
this is the intended behaviour or not, but I figured it'd be a good
thing if using the Docker container everything worked out of the box.