fixes for issues  (#20)
* basic_tutorial runthrough, improvement suggestions

I imposed a couple of conventions:

- column and row names in the tables shown are in the text written in
**bold**
- variable names, parameters, and code in general is formatted as `code`
- files/names in the Zarr directory are formatted as 'cellData'

* Dockerfile, tzdata

The existing tzdata solution did not work for me when building the
image, with these updates it worked for me

* Dockerfile, dask methods working

Scarf method `Datastore.plot_marker_heatmap` did not work without the
installed package fsspec

* Dockerfile, automatic scarf installation

I added automatic installation of the local scarf module. Not sure if
this is the intended behaviour or not, but I figured it'd be a good
thing if using the Docker container everything worked out of the box.

* solves issue #17

* Dockerfile, basic vignette works out-of-box now

- removed redundant definition of TZ variable
- added the project bin dir to PATH, so that scarf can find and use that
for the run_tsne method
	- solves #18
- added a build of jupyter lab so that user does not have to do that at
startup (got a prompt for build required after installing jupytext)

* basic vignette, removed hardcoded tree graph info

* docs, automatic build of notebooks

the change in conf.py from
`nbshpinx_execute = 'never'`
to
`nbshpinx_execute = 'auto'`
was the key. With 'auto' the notebooks are executed if there are _no
output_ cells present in them.
https://nbsphinx.readthedocs.io/en/0.8.0/never-execute.html

* fixed format in docstring that disabled API docs build