working towards fuller docstring coverage (atm 50%) (#25)
* a few grammatical changes to docs

* started on docstrings

Started adding missing docstrings. Marking where to add docstrings with
TODOs, and started filling in some. Am generally trying to add
docstrings for public-changing parts (such as for the module, for
exported classes and methods etc).

* further along in raising docstring coverage

WIP

* merged parashar's latest commits

* moved module docstrings

moved module docstrings so they actually show up when requested in REPL
(when they are not at the top of the file they are not shown)

* more docstrings