docstrings, uniformity and sphinx acrobatics
1. Circumvented sphinx's parsing of "Methods:" for the module-level
docstrings, this way we can get a listing of the most important methods
and classes in a module, without having several method/class
documentations in different places on the API docs page.
2. Small changes to make the docstrings more .. (Mostly capitalise start
of sentences and adding periods)
3. Added a few `automodule` sphinx directives, as discussed in #34 . We
can remove or add more of these after a conclusion of the discussion
there.