solved numpy error that gave test fails
There was an error resulting from numpy making a change in their C usage
(see  ), this made our paris clustering test fail and subsequently a lot
of tests depending on there being clusters in the datastore object. If I
with python 3.9 install from the current requirements file the tests are
good again.